sw_services: xilloader: Fixing MISRA C Rule 10.4 in xilloader
MISRA C Rule 10.4 states: Both operands of an operator in which the usual arithmetic conversions are performed shall have the same essential type category

This patch fixes MISRA C Rule 10.4 in xilloader by casting TRUE(1U) & FALSE(0U) to u8 because that MISRA C confuses TRUE/FALSE as boolean type


Acked-for-series: Vikram Sreenivasa Batchali <b.vikram@xilinx.com>

+++ lib/sw_services/xilloader/src/xloader.c
@@ -423,7 +423,7 @@ static int XLoader_ReadAndValidateHdrs(XilPdi* PdiPtr, u32 RegVal)
 			(PdiPtr->PdiSrc == XLOADER_PDI_SRC_EMMC_RAW_BP2)) {
 
 			PdiPtr->MetaHdr.FlashOfstAddr = PdiPtr->PdiAddr + \
-				(RegVal * XLOADER_IMAGE_SEARCH_OFFSET);
+				((u64)RegVal * XLOADER_IMAGE_SEARCH_OFFSET);
 			if ((PdiPtr->PdiSrc == XLOADER_PDI_SRC_QSPI24) || \
 				(PdiPtr->PdiSrc == XLOADER_PDI_SRC_QSPI32)) {
 #ifdef XLOADER_QSPI
@@ -1038,7 +1038,7 @@ int XLoader_LoadImage(XilPdi *PdiPtr)
 		goto END;
 	}
 
-	if ((PdiPtr->DelayLoad == FALSE) &&
+	if ((PdiPtr->DelayLoad == (u8)FALSE) &&
 		(PdiPtr->MetaHdr.ImgHdr[PdiPtr->ImageNum].ImgID !=
 		XPLMI_INVALID_IMG_ID)) {
 		Status = XPlmi_StoreImageInfo(&ImageInfo);
